+++ src/GpuCache.h
@@ -34,46 +34,46 @@ typedef boost::shared_lock<gpu_cache_rwmutex> gpu_cache_readLock;
 typedef boost::unique_lock<gpu_cache_rwmutex> gpu_cache_writeLock;
 
 
-extern void mog_malloc_gpu(int32_t device_id,
-                           int32_t slab_size,
+extern void mog_malloc_gpu(int64_t device_id,
+                           int64_t slab_size,
                            char** slabs,
-                           int32_t slab_id);
+                           int64_t slab_id);
 
-extern void mog_memcpy_cpu_to_gpu(int32_t device_id, char* dst, const char* src, int32_t page_size);
+extern void mog_memcpy_cpu_to_gpu(int64_t device_id, char* dst, const char* src, int64_t page_size);
 
-extern void mog_memcpy_gpu_to_cpu(int32_t device_id, char* dst, const char* src, int32_t page_size);
+extern void mog_memcpy_gpu_to_cpu(int64_t device_id, char* dst, const char* src, int64_t page_size);
 
-extern void mog_memcpy_gpu_to_gpu(int32_t device_id, char* dst, const char* src, int32_t slabe_size);
+extern void mog_memcpy_gpu_to_gpu(int64_t device_id, char* dst, const char* src, int64_t slabe_size);
 
-extern void mog_free_gpu(int32_t device_id, char* data_p);
+extern void mog_free_gpu(int64_t device_id, char* data_p);
 
 namespace cap {
 
 struct GPUCachedItem {
-    //int32_t key;
-    int32_t length;
-    int32_t length_type;
-    std::vector<int32_t> slabs;
+    //int64_t key;
+    int64_t length;
+    int64_t length_type;
+    std::vector<int64_t> slabs;
 };
 
 class GpuCache {
 public:
     GpuCache();
-    GpuCache(const int32_t device_id);
+    GpuCache(const int64_t device_id);
     virtual ~GpuCache();
 
-    int32_t device_id;
+    int64_t device_id;
 
     gpu_cache_rwmutex rw_cache_lock;
-    int32_t page_size;
-    int32_t slab_num;
-    int32_t page_num;
-    std::atomic<int32_t> free_slab_num;
+    int64_t page_size;
+    int64_t slab_num;
+    int64_t page_num;
+    std::atomic<int64_t> free_slab_num;
     //std::mutex cached_item_queue_lock;
 
-    std::queue<int32_t> free_slabs;
+    std::queue<int64_t> free_slabs;
 
-    std::atomic<int32_t> read_thread_num;
+    std::atomic<int64_t> read_thread_num;
     char cpu_buffer[10][1024*1024];
 
 
@@ -84,18 +84,18 @@ public:
     char** grouped_slabs;
 
 
-    int32_t initial_flag = 0;
-    static const int32_t slab_size[11];//KB
+    int64_t initial_flag = 0;
+    static const int64_t slab_size[11];//KB
 
-    int32_t initial(const int32_t page_size, const int32_t page_num);
-    int32_t allocate_memory();
+    int64_t initial(const int64_t page_size, const int64_t page_num);
+    int64_t allocate_memory();
 
-    int32_t close();
+    int64_t close();
 
-    int32_t read(const std::string& key, char *value);
-    int32_t insert(const std::string &key,
-                   const int32_t length,
-                   const int32_t length_type,
+    int64_t read(const std::string& key, char *value);
+    int64_t insert(const std::string &key,
+                   const int64_t length,
+                   const int64_t length_type,
                    char *value);
 };
