+++ src/WriteBuffer.h
@@ -38,10 +38,10 @@ typedef boost::unique_lock<ssd_cache_rwmutex> wbuffer_writeLock;
 struct kv_info {
     std::string key;
     char* value;
-    int32_t value_length;
-    int32_t value_length_type;
-//    int32_t w_lock;
-//    int32_t r_lock;
+    int64_t value_length;
+    int64_t value_length_type;
+//    int64_t w_lock;
+//    int64_t r_lock;
 };
 
 class WriteBuffer {
@@ -49,14 +49,14 @@ public:
     WriteBuffer();
     virtual ~WriteBuffer();
 
-    int32_t page_size;
-    int32_t page_per_block;
-    int32_t page_num;
-    int32_t max_page_num;
-    int32_t initial_flag;
-    int32_t flush_thread_num;
-    int32_t latest_block_id;
-    static const int32_t slab_size[11];//KB
+    int64_t page_size;
+    int64_t page_per_block;
+    int64_t page_num;
+    int64_t max_page_num;
+    int64_t initial_flag;
+    int64_t flush_thread_num;
+    int64_t latest_block_id;
+    static const int64_t slab_size[11];//KB
 
     std::string ssd_path;
     std::string dfs_path;
@@ -67,26 +67,26 @@ public:
 
     std::queue<size_t> kv_list;
     std::unordered_map<size_t, kv_info> kv_store;
-    std::unordered_map<int32_t, std::vector<kv_info>> blocks;
-    std::atomic<int32_t> total_size_without_block;
-    std::atomic<int32_t> total_size;
+    std::unordered_map<int64_t, std::vector<kv_info>> blocks;
+    std::atomic<int64_t> total_size_without_block;
+    std::atomic<int64_t> total_size;
     boost::threadpool::pool *flush_thread_pool;
 
 //    std::vector<>
 
     //std::vector<std::vector<char*>> buffer;
 
-    int32_t initial(const int32_t page_size,
-                const int32_t page_per_block,
-                const int32_t page_num);
-    int32_t allocate_memory();
+    int64_t initial(const int64_t page_size,
+                const int64_t page_per_block,
+                const int64_t page_num);
+    int64_t allocate_memory();
 
-    int32_t write(const std::string& key, const char *value, int32_t length);
-    int32_t read(const std::string& key, char *value);
+    int64_t write(const std::string& key, const char *value, int64_t length);
+    int64_t read(const std::string& key, char *value);
     static WriteBuffer& singleton();
-    static void flush(int32_t block_id);
-    int32_t sync();
-    int32_t close();
+    static void flush(int64_t block_id);
+    int64_t sync();
+    int64_t close();
 };
 
 } /* namespace cap */
