+++ src/mog_GPU.cu
@@ -13,50 +13,50 @@ static void HandleError( cudaError_t err,
 }
 #define HANDLE_ERROR( err ) (HandleError( err, __FILE__, __LINE__ ))
 
-void mog_malloc_gpu(int32_t device_id, 
-				int32_t slab_size, 
+void mog_malloc_gpu(int64_t device_id, 
+				int64_t slab_size, 
 				char** slabs,
-				int32_t slab_id) {
+				int64_t slab_id) {
 	cudaSetDevice(device_id);
 	char *new_slab;
 	HANDLE_ERROR(cudaMalloc((void **)&new_slab, slab_size));
 	slabs[slab_id] = new_slab;
 }
 
-char* mog_malloc_gpu(int32_t device_id, 
-				int32_t size) {
+char* mog_malloc_gpu(int64_t device_id, 
+				int64_t size) {
 	cudaSetDevice(device_id);
 	char *data_p;
 	HANDLE_ERROR(cudaMalloc((void **)&data_p, size));
 	return data_p;
 }
 
-void mog_free_gpu(int32_t device_id, 
+void mog_free_gpu(int64_t device_id, 
 				char* data_p) {
 	cudaSetDevice(device_id);
 	cudaFree ((void*) data_p);
 }
 
-void mog_memcpy_cpu_to_gpu(int32_t device_id, char* dst, const char* src, int32_t slabe_size) {
+void mog_memcpy_cpu_to_gpu(int64_t device_id, char* dst, const char* src, int64_t slabe_size) {
 	cudaSetDevice(device_id);
 	HANDLE_ERROR(cudaMemcpy(dst, src, slabe_size, cudaMemcpyHostToDevice));
 	//cudaDeviceSynchronize();
 }
 
-void mog_memcpy_gpu_to_cpu(int32_t device_id, char* dst, const char* src, int32_t slabe_size) {
+void mog_memcpy_gpu_to_cpu(int64_t device_id, char* dst, const char* src, int64_t slabe_size) {
 	cudaSetDevice(device_id);
 	HANDLE_ERROR(cudaMemcpy(dst, src, slabe_size, cudaMemcpyDeviceToHost));
 	//cudaDeviceSynchronize();
 }
 
-void mog_memcpy_gpu_to_gpu(int32_t device_id, char* dst, const char* src, int32_t slabe_size) {
+void mog_memcpy_gpu_to_gpu(int64_t device_id, char* dst, const char* src, int64_t slabe_size) {
 	cudaSetDevice(device_id);
 	HANDLE_ERROR(cudaMemcpy(dst, src, slabe_size, cudaMemcpyDeviceToDevice));
 	//cudaDeviceSynchronize();
 }
 
 /*
-void mog_vectorAdd(int32_t device_id, const char *A, const char *B, char *C, int numElements) {
+void mog_vectorAdd(int64_t device_id, const char *A, const char *B, char *C, int numElements) {
 	cudaError_t err = cudaSuccess;
 	
 	cudaSetDevice(device_id);
